Move common parser setup to ramutils.cli
